LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/relevant-mutations-expected.txt


PASS src set 
FAIL src changed assert_unreached: update the image data didn't run Reached unreachable code
PASS src removed 
PASS srcset set 
FAIL srcset changed assert_unreached: update the image data didn't run Reached unreachable code
PASS srcset removed 
PASS sizes set 
PASS sizes changed 
PASS sizes removed 
FAIL src set to same value assert_unreached: update the image data didn't run Reached unreachable code
PASS crossorigin absent to empty, src absent 
PASS crossorigin absent to anonymous, src absent 
PASS crossorigin absent to use-credentials, src absent 
PASS crossorigin empty to absent, src absent 
PASS crossorigin empty to use-credentials, src absent 
PASS crossorigin anonymous to absent, src absent 
PASS crossorigin anonymous to use-credentials, src absent 
PASS crossorigin use-credentials to absent, src absent 
PASS crossorigin use-credentials to empty, src absent 
PASS crossorigin use-credentials to anonymous, src absent 
PASS crossorigin use-credentials to invalid, src absent 
FAIL crossorigin absent to empty, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin absent to anonymous, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin absent to use-credentials, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin empty to absent, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin empty to use-credentials, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin anonymous to absent, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin anonymous to use-credentials, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin use-credentials to absent, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin use-credentials to empty, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin use-credentials to anonymous, src already set assert_unreached: update the image data didn't run Reached unreachable code
FAIL crossorigin use-credentials to invalid, src already set assert_unreached: update the image data didn't run Reached unreachable code
PASS referrerpolicy absent to no-referrer-when-downgrade, src absent 
PASS referrerpolicy absent to no-referrer, src absent 
PASS referrerpolicy empty to no-referrer-when-downgrade, src absent 
PASS referrerpolicy empty to no-referrer, src absent 
PASS referrerpolicy no-referrer-when-downgrade to absent, src absent 
PASS referrerpolicy no-referrer-when-downgrade to empty, src absent 
PASS referrerpolicy no-referrer-when-downgrade to no-referrer, src absent 
PASS referrerpolicy no-referrer-when-downgrade to invalid, src absent 
PASS referrerpolicy no-referrer to absent, src absent 
PASS referrerpolicy no-referrer to empty, src absent 
PASS referrerpolicy no-referrer to no-referrer-when-downgrade, src absent 
PASS referrerpolicy no-referrer to invalid, src absent 
PASS referrerpolicy absent to no-referrer-when-downgrade, src already set 
PASS referrerpolicy absent to no-referrer, src already set 
PASS referrerpolicy empty to no-referrer-when-downgrade, src already set 
PASS referrerpolicy empty to no-referrer, src already set 
PASS referrerpolicy no-referrer-when-downgrade to absent, src already set 
PASS referrerpolicy no-referrer-when-downgrade to empty, src already set 
PASS referrerpolicy no-referrer-when-downgrade to no-referrer, src already set 
PASS referrerpolicy no-referrer-when-downgrade to invalid, src already set 
PASS referrerpolicy no-referrer to absent, src already set 
PASS referrerpolicy no-referrer to empty, src already set 
PASS referrerpolicy no-referrer to no-referrer-when-downgrade, src already set 
PASS referrerpolicy no-referrer to invalid, src already set 
FAIL inserted into picture assert_unreached: update the image data didn't run Reached unreachable code
FAIL removed from picture assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source inserted assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source removed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has srcset set assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has srcset changed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has srcset removed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has sizes set assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has sizes changed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has sizes removed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has media set assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has media changed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has media removed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has type set assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has type changed assert_unreached: update the image data didn't run Reached unreachable code
FAIL parent is picture, previous source has type removed assert_unreached: update the image data didn't run Reached unreachable code
PASS srcset is set to same value 
PASS sizes is set to same value 
PASS crossorigin state not changed: absent, removeAttribute 
PASS crossorigin state not changed: empty to anonymous 
PASS crossorigin state not changed: anonymous to foobar 
PASS crossorigin state not changed: use-credentials to USE-CREDENTIALS 
PASS referrerpolicy state not changed: absent, removeAttribute 
PASS referrerpolicy state not changed: empty to empty 
PASS referrerpolicy state not changed: empty to invalid 
PASS referrerpolicy state not changed: absent to invalid 
PASS referrerpolicy state not changed: no-referrer to NO-REFERRER 
PASS referrerpolicy state not changed: invalid to other-invalid 
PASS inserted into picture ancestor 
PASS removed from picture ancestor 
PASS ancestor picture has a source inserted 
PASS ancestor picture has a source removed 
PASS ancestor picture; previous sibling source inserted 
PASS ancestor picture; previous sibling source removed 
PASS parent is picture, following sibling source inserted 
PASS parent is picture, following sibling source removed 
PASS parent is picture, following sibling source has srcset set 
PASS media on img set 
PASS type on img set 
PASS class on img set 
PASS alt on img set 
PASS src on previous sibling source set 
PASS class on previous sibling source set 
PASS inserted/removed children of img 
PASS picture is inserted; img has src 
PASS picture is inserted; img has srcset 
PASS picture is inserted; img has previous sibling source 
PASS picture is inserted; img has following sibling source 
PASS picture is removed; img has src 
PASS picture is removed; img has srcset 
PASS picture is removed; img has previous sibling source 
PASS picture is removed; img has following sibling source 
PASS parent is picture, following img inserted 
PASS parent is picture, following img removed 
PASS parent is picture, following img has src set 
PASS parent is picture, following img has srcset set 
PASS parent is picture, following img has sizes set 
                                                                                                                                     
